refactor: address opus review follow-ups for issue 132
