+++ src/validate.ts
@@ -407,276 +407,17 @@ export async function validateStep(args: { rootDir: string; checkpoint: Checkpoi
   if (args.step.kind === "volume") {
     const volume = args.checkpoint.current_volume;
     const range = await resolveVolumeChapterRange({ rootDir: args.rootDir, current_volume: volume, last_completed_chapter: args.checkpoint.last_completed_chapter });
-    const rels = volumeStagingRelPaths(volume);
-
-    const requireVolumePlanArtifacts = async (): Promise<void> => {
-      requireFile(await pathExists(join(args.rootDir, rels.outlineMd)), rels.outlineMd);
-      requireFile(await pathExists(join(args.rootDir, rels.storylineScheduleJson)), rels.storylineScheduleJson);
-      requireFile(await pathExists(join(args.rootDir, rels.foreshadowingJson)), rels.foreshadowingJson);
-      requireFile(await pathExists(join(args.rootDir, rels.newCharactersJson)), rels.newCharactersJson);
-      requireFile(await pathExists(join(args.rootDir, rels.chapterContractsDir)), rels.chapterContractsDir);
-      for (let ch = range.start; ch <= range.end; ch++) {
-        requireFile(await pathExists(join(args.rootDir, rels.chapterContractJson(ch))), rels.chapterContractJson(ch));
-      }
-    };
-
-    const outlineAbs = join(args.rootDir, rels.outlineMd);
-    const parseOutlineChapterMetadata = async (): Promise<Map<number, { storylineId: string; excitementType: ExcitementType | null }>> => {
-      const text = await readTextFile(outlineAbs);
-      if (text.trim().length === 0) throw new NovelCliError(`Empty outline file: ${rels.outlineMd}`, 2);
-
-      const lines = text.split(/\r?\n/u);
-      const chapters: Array<{ chapter: number; startLine: number; endLine: number }> = [];
-      const chapterHeadingRe = /^###\s*第\s*(\d+)\s*章/u;
-      for (let i = 0; i < lines.length; i++) {
-        const m = chapterHeadingRe.exec(lines[i] ?? "");
-        if (!m) continue;
-        const chapter = Number.parseInt(m[1] ?? "", 10);
-        if (!Number.isInteger(chapter) || chapter < 1) continue;
-        chapters.push({ chapter, startLine: i, endLine: lines.length });
-      }
-      if (chapters.length === 0) {
-        throw new NovelCliError(
-          `Invalid outline: could not parse any chapter headings. Expected lines like "### 第 1 章" (spaces are flexible). File: ${rels.outlineMd}`,
-          2
-        );
-      }
-      for (let i = 0; i < chapters.length; i++) {
-        const cur = chapters[i]!;
-        const next = chapters[i + 1];
-        if (next) cur.endLine = next.startLine;
-      }
-
-      const seen = new Set<number>();
-      for (const c of chapters) {
-        if (seen.has(c.chapter)) throw new NovelCliError(`Invalid outline: duplicate chapter block for chapter ${c.chapter} (${rels.outlineMd}).`, 2);
-        seen.add(c.chapter);
-      }
-      for (let ch = range.start; ch <= range.end; ch++) {
-        if (!seen.has(ch)) {
-          throw new NovelCliError(
-            `Invalid outline: missing chapter block for chapter ${ch} (expected continuous coverage ${range.start}-${range.end}). File: ${rels.outlineMd}`,
-            2
-          );
-        }
-      }
-      assertOutlineHasNoOutOfRangeChapters({ seen, relPath: rels.outlineMd, range });
-
-      const requiredKeys = [
-        "Storyline",
-        "POV",
-        "Location",
-        "Conflict",
-        "Arc",
-        "Foreshadowing",
-        "StateChanges",
-        "TransitionHint"
-      ] as const;
-      const excitementPrefix = "- **ExcitementType**:";
-
-      const chapterMetadata = new Map<number, { storylineId: string; excitementType: ExcitementType | null }>();
-      for (const c of chapters) {
-        if (c.chapter < range.start || c.chapter > range.end) continue;
-        const blockLines = lines.slice(c.startLine, c.endLine);
-        const keyFound = new Set<string>();
-        let storylineId: string | null = null;
-        let excitementType: ExcitementType | null = null;
-        for (const line of blockLines) {
-          for (const k of requiredKeys) {
-            const prefix = `- **${k}**:`;
-            if (line.startsWith(prefix)) {
-              keyFound.add(k);
-              if (k === "Storyline") {
-                const val = line.slice(prefix.length).trim();
-                if (val.length > 0) storylineId = val;
-              }
-            }
-          }
-
-          if (line.startsWith(excitementPrefix)) {
-            const rawExcitementType = line.slice(excitementPrefix.length).trim();
-            const normalizedExcitementType = normalizeExcitementType(rawExcitementType);
-            if (normalizedExcitementType === undefined) {
-              warnings.push(
-                `Outline chapter ${c.chapter} has unknown ExcitementType '${rawExcitementType}' in ${rels.outlineMd}; treating as null.`
-              );
-            } else {
-              excitementType = normalizedExcitementType;
-            }
-          }
-        }
-        for (const k of requiredKeys) {
-          if (!keyFound.has(k)) {
-            throw new NovelCliError(
-              `Invalid outline: chapter ${c.chapter} block missing required key '${k}' line. File: ${rels.outlineMd}`,
-              2
-            );
-          }
-        }
-        if (!storylineId) {
-          throw new NovelCliError(`Invalid outline: chapter ${c.chapter} missing non-empty Storyline value. File: ${rels.outlineMd}`, 2);
-        }
-        chapterMetadata.set(c.chapter, { storylineId, excitementType });
-      }
-
-      return chapterMetadata;
-    };
-
-    const validateNewCharacters = (raw: unknown): void => {
-      if (!Array.isArray(raw)) throw new NovelCliError(`Invalid ${rels.newCharactersJson}: expected an array.`, 2);
-      for (const [idx, item] of raw.entries()) {
-        if (!isPlainObject(item)) throw new NovelCliError(`Invalid ${rels.newCharactersJson}: entry ${idx} must be an object.`, 2);
-        const obj = item as Record<string, unknown>;
-        const name = typeof obj.name === "string" ? obj.name.trim() : "";
-        const role = typeof obj.role === "string" ? obj.role.trim() : "";
-        const brief = typeof obj.brief === "string" ? obj.brief.trim() : "";
-        const firstChapter = typeof obj.first_chapter === "number" && Number.isInteger(obj.first_chapter) ? obj.first_chapter : null;
-        if (name.length === 0) throw new NovelCliError(`Invalid ${rels.newCharactersJson}: entry ${idx} missing name.`, 2);
-        if (brief.length === 0) throw new NovelCliError(`Invalid ${rels.newCharactersJson}: entry ${idx} missing brief.`, 2);
-        if (firstChapter === null || firstChapter < range.start || firstChapter > range.end) {
-          throw new NovelCliError(
-            `Invalid ${rels.newCharactersJson}: entry ${idx} first_chapter=${String(obj.first_chapter)} out of range (${range.start}-${range.end}).`,
-            2
-          );
-        }
-        if (role !== "antagonist" && role !== "supporting" && role !== "minor") {
-          throw new NovelCliError(
-            `Invalid ${rels.newCharactersJson}: entry ${idx} role must be one of: antagonist, supporting, minor.`,
-            2
-          );
-        }
-      }
-    };
-
-    const validateSchedule = (raw: unknown, chapterMetadata: Map<number, { storylineId: string; excitementType: ExcitementType | null }>): void => {
-      if (!isPlainObject(raw)) throw new NovelCliError(`Invalid ${rels.storylineScheduleJson}: must be an object.`, 2);
-      const obj = raw as Record<string, unknown>;
-      const active = obj.active_storylines;
-      if (!Array.isArray(active)) throw new NovelCliError(`Invalid ${rels.storylineScheduleJson}: missing active_storylines array.`, 2);
-      const activeIds: string[] = [];
-      for (const [idx, v] of active.entries()) {
-        if (typeof v !== "string") {
-          warnings.push(`WARN ${rels.storylineScheduleJson}: active_storylines[${idx}] is not a string; ignoring.`);
-          continue;
-        }
-        const id = v.trim();
-        if (id.length > 0) activeIds.push(id);
-      }
-      if (activeIds.length === 0) throw new NovelCliError(`Invalid ${rels.storylineScheduleJson}: active_storylines must be non-empty.`, 2);
-      if (activeIds.length > 4) throw new NovelCliError(`Invalid ${rels.storylineScheduleJson}: active_storylines length must be <= 4.`, 2);
-      const activeSet = new Set(activeIds);
-
-      for (const [ch, metadata] of chapterMetadata.entries()) {
-        const storylineId = metadata.storylineId;
-        if (!activeSet.has(storylineId)) {
-          throw new NovelCliError(
-            `Invalid volume plan: outline chapter ${ch} Storyline=${storylineId} not in storyline-schedule.json.active_storylines.`,
-            2
-          );
-        }
-      }
-    };
-
-    const validateForeshadowingPlan = (raw: unknown): void => {
-      if (!isPlainObject(raw)) {
-        throw new NovelCliError(`Invalid ${rels.foreshadowingJson}: expected JSON object with {schema_version, items}.`, 2);
-      }
-      const obj = raw as Record<string, unknown>;
-      const schema = obj.schema_version;
-      if (schema !== 1) warnings.push(`Unexpected schema_version in ${rels.foreshadowingJson}.`);
-      const items = obj.items;
-      if (!Array.isArray(items)) throw new NovelCliError(`Invalid ${rels.foreshadowingJson}: items must be an array.`, 2);
-    };
-
-    const validateContracts = async (chapterMetadata: Map<number, { storylineId: string; excitementType: ExcitementType | null }>): Promise<void> => {
-      await assertContractDirMatchesRange({ rootDir: args.rootDir, rels, range });
-      for (let ch = range.start; ch <= range.end; ch++) {
-        const rel = rels.chapterContractJson(ch);
-        const raw = await readJsonFile(join(args.rootDir, rel));
-        if (!isPlainObject(raw)) throw new NovelCliError(`Invalid ${rel}: must be an object.`, 2);
-        const obj = raw as Record<string, unknown>;
-        const chapter = requireNumberField(obj, "chapter", rel);
-        if (!Number.isInteger(chapter) || chapter !== ch) throw new NovelCliError(`Invalid ${rel}: chapter=${chapter} expected ${ch}.`, 2);
-        const storylineId = requireStringField(obj, "storyline_id", rel).trim();
-        const expectedStorylineId = chapterMetadata.get(ch)?.storylineId ?? null;
-        if (!expectedStorylineId || storylineId !== expectedStorylineId) {
-          throw new NovelCliError(
-            `Invalid ${rel}: storyline_id=${storylineId} expected ${expectedStorylineId ?? "(missing in outline)"}.`,
-            2
-          );
-        }
-
-        if (!Object.prototype.hasOwnProperty.call(obj, "excitement_type")) {
-          warnings.push(`Missing optional excitement_type in ${rel}; treating as null.`);
-        } else {
-          const normalizedExcitementType = normalizeExcitementType(obj.excitement_type);
-          if (normalizedExcitementType === undefined) {
-            warnings.push(`Unknown excitement_type '${String(obj.excitement_type)}' in ${rel}; treating as null.`);
-          }
-        }
-
-        const objectivesRaw = obj.objectives;
-        if (!Array.isArray(objectivesRaw)) throw new NovelCliError(`Invalid ${rel}: objectives must be an array.`, 2);
-        const hasRequired = objectivesRaw.some((it) => isPlainObject(it) && (it as Record<string, unknown>).required === true);
-        if (!hasRequired) throw new NovelCliError(`Invalid ${rel}: objectives must include at least one entry with required=true.`, 2);
-
-        // Chain propagation (minimal): prev.postconditions.state_changes keys must exist in current.preconditions.character_states.
-        const prevChapter = ch - 1;
-        if (prevChapter >= 1) {
-          const prevRel =
-            prevChapter >= range.start
-              ? rels.chapterContractJson(prevChapter)
-              : volumeFinalRelPaths(volumeForChapter(prevChapter)).chapterContractJson(prevChapter);
-
-          if (await pathExists(join(args.rootDir, prevRel))) {
-            const prevRaw = await readJsonFile(join(args.rootDir, prevRel));
-            if (isPlainObject(prevRaw)) {
-              const prevObj = prevRaw as Record<string, unknown>;
-              const post = isPlainObject(prevObj.postconditions) ? (prevObj.postconditions as Record<string, unknown>) : null;
-              const stateChanges = post && isPlainObject(post.state_changes) ? (post.state_changes as Record<string, unknown>) : null;
-              const pre = isPlainObject(obj.preconditions) ? (obj.preconditions as Record<string, unknown>) : null;
-              const characterStates = pre && isPlainObject(pre.character_states) ? (pre.character_states as Record<string, unknown>) : null;
-
-              if (stateChanges && Object.keys(stateChanges).length > 0) {
-                if (!characterStates) {
-                  throw new NovelCliError(
-                    `Invalid ${rel}: missing preconditions.character_states required by chain propagation from ${prevRel}.`,
-                    2
-                  );
-                }
-                for (const characterKey of Object.keys(stateChanges)) {
-                  if (!(characterKey in characterStates)) {
-                    throw new NovelCliError(
-                      `Invalid ${rel}: preconditions.character_states missing '${characterKey}' (required by ${prevRel}.postconditions.state_changes).`,
-                      2
-                    );
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    };
-
     if (args.step.phase === "commit") {
       throw new NovelCliError(`Use 'novel commit --volume ${volume}' for commit.`, 2);
     }
 
-    await requireVolumePlanArtifacts();
-
-    const chapterMetadata = await parseOutlineChapterMetadata();
-
-    const scheduleRaw = await readJsonFile(join(args.rootDir, rels.storylineScheduleJson));
-    validateSchedule(scheduleRaw, chapterMetadata);
-
-    // JSON existence + basic schema.
-    const foreshadowingRaw = await readJsonFile(join(args.rootDir, rels.foreshadowingJson));
-    validateForeshadowingPlan(foreshadowingRaw);
-    const newCharsRaw = await readJsonFile(join(args.rootDir, rels.newCharactersJson));
-    validateNewCharacters(newCharsRaw);
-
-    await validateContracts(chapterMetadata);
+    await validateVolumePlanArtifacts({
+      rootDir: args.rootDir,
+      volume,
+      range,
+      rels: volumeStagingRelPaths(volume),
+      warnings
+    });
 
     return { ok: true, step: stepId, warnings };
   }
